Move all indexers to own folder
