+++ src/cita/itemWrapper.ts
@@ -1,8 +1,8 @@
-import Crossref from "./crossref";
-import OpenAlex from "./openalex";
-import OpenCitations from "./opencitations";
+import Crossref from "./indexers/crossref";
+import OpenAlex from "./indexers/openalex";
+import OpenCitations from "./indexers/opencitations";
 import Progress from "./progress";
-import Semantic from "./semantic";
+import Semantic from "./indexers/semantic";
 import Wikicite from "./wikicite";
 import Wikidata from "./wikidata";
 import PID from "./PID";
